added dummy components for manage friends page
