+++ api_dev/dropin-frontend/src/components/AddFriend/AddFriend.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function AddFriend(){
+    return(
+        <h2>Add Friend:</h2>
+    )
+}
