+++ api_dev/dropin-frontend/src/components/PendingRequests/PedingRequests.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function PendingRequests(){
+    return(
+        <h2>Pending Requests:</h2>
+    )
+}
